+++ apps/ai-dial-admin/src/constants/i18n.ts
@@ -1234,6 +1234,7 @@ export enum ContainersI18nKey {
   ContainerSuccessDescription = 'Containers.ContainerSuccessDescription',
   Pod = 'Containers.Pod',
   Pods = 'Containers.Pods',
+  ModelDetails = 'Containers.ModelDetails',
   ContainerRestartWarning = 'Containers.ContainerRestartWarning',
   ContainersRestartWarning = 'Containers.ContainersRestartWarning',
   ShowRelatedContainers = 'Containers.ShowRelatedContainers',
