+++ apps/ai-dial-admin/src/locales/en.ts
@@ -1254,6 +1254,7 @@ export default {
     Stopping: 'Stopping...',
     Pod: 'Pod',
     Pods: 'Pods',
+    ModelDetails: 'Model details',
     ContainerRestartWarning: '{entityType} will be restarted after changing image version.',
     ContainersRestartWarning: '{entityType} with "Running" status will be restarted.',
     ShowRelatedContainers: 'Show containers related to this image',
